+++ sandbox/__init__.py
@@ -1,0 +1,1 @@
+import sandbox._log
