--- sandbox/logging_config_ex.json
@@ -1,35 +1,0 @@
-{
-    "version": 1,
-    "disable_existing_loggers": false,
-    "propagate": true,
-    "root": {
-        "": {
-            "handlers": [
-                "console"
-            ],
-            "level": "NOTSET"
-        }
-    },
-    "formatters": {
-        "default": {
-            "format": "[%(asctime)s] [%(levelname)s] > [%(filename)s:%(lineno)d] [%(funcName)s]: %(message)s",
-            "datefmt": "%Y-%m-%d %H:%M:%S"
-        }
-    },
-    "handlers": {
-        "console": {
-            "class": "logging.StreamHandler",
-            "formatter": "default",
-            "level": "DEBUG",
-            "stream": "ext://sys.stdout"
-        }
-    },
-    "loggers": {
-        "": {
-            "handlers": [
-                "console"
-            ],
-            "level": "NOTSET"
-        }
-    }
-}
