--- src.bak/red_log/configs/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-from ._configs import DEFAULT_LOGGING_CONFIG, CONFIG_DICT_BASE
