--- src.bak/red_log/core/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import annotations
-
-from ..constants import DEFAULT_DATE_FMT, DEFAULT_FMT, DEFAULT_LOGGER_STYLE
-from ..constants import LogLevelsEnum
