--- src.bak/red_log/formatters/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import annotations
-
-from ._formatters import DEFAULT_LOG_FORMATTER, LEVEL_MSG, TS_MSG
-from ._formatters import get_formatter
