--- src.bak/red_log/handlers/__init__.py
@@ -1,14 +1,0 @@
-from __future__ import annotations
-
-# from .prefab import (
-#     APP_FILE_HANDLER,
-#     CLI_CONSOLE_HANDLER,
-#     CONSOLE_HANDLER,
-#     ERR_FILE_HANDLER,
-#     ROTATING_FILE_HANDLER,
-#     TIMED_ROTATING_FILE_HANDLER,
-# )
-from . import prefab
-from ._handlers import get_handler, get_handler_config
-from .validators import VALID_HANDLER_CLASSES
-from ._types import HandlerType
