--- src.bak/red_log/handlers/prefab/__init__.py
@@ -1,9 +1,0 @@
-from ._preconfigured_handlers import (
-    CONSOLE_HANDLER,
-    APP_FILE_HANDLER,
-    ERR_FILE_HANDLER,
-    CLI_CONSOLE_HANDLER,
-    ROTATING_ERR_FILE_HANDLER,
-    ROTATING_FILE_HANDLER,
-    TIMED_ROTATING_FILE_HANDLER,
-)
