--- src.bak/red_log/loggers/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from . import third_party
-from ._loggers import DEFAULT_ROOT_LOGGER, SILENT_LOGGER
-from .third_party import REQUESTS_LOGGER, UVICORN_LOGGER
