--- src.bak/red_log/utils/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-from ._methods import configure_root_logger, test_log_messages
