--- src.bak/red_log/utils/gen_config/methods.py
@@ -1,2 +1,0 @@
-import logging
-import logging.config
