+++ src/red_log/__init__.py
@@ -1,15 +1,16 @@
+import logging
+
 from . import constants
-
-# from .config_dicts import BASE_LOGGING_CONFIG_DICT
-
-# from . import formatters, handlers, loggers
-
-# from .formatters import FormatterFactory
-# from .handlers import HandlerFactory
-# from .loggers import LoggerFactory
 
 from . import base
 from . import filters
 from . import formatters
 from . import handlers
 from . import configs
+from . import loggers
+
+from .configs import LoggingConfig
+
+rootlogger = logging.getLogger("red_log")
+if rootlogger.level == logging.NOTSET:
+    rootlogger.setLevel(logging.WARN)
