+++ src/red_log/constants.py
@@ -1,10 +1,10 @@
 DEFAULT_FMT: str = (
     "[%(asctime)s] [%(levelname)s] > [%(filename)s:%(lineno)d] [%(funcName)s]: %(message)s"
 )
-DEFAULT_SIMPLE_FMT: str = (
+SIMPLE_FMT: str = (
     "[%(asctime)s] [%(levelname)s] > [%(filename)s:%(lineno)d]: %(message)s"
 )
-DEFAULT_DETAIL_FMT: str = (
+DETAIL_FMT: str = (
     "[%(astime)s] [Logger:%(name)s] [pid:%(process)d] [procname:%(processName)s] [%(levelname)s] [module:%(module)s] > [%(filename)s.%(funcname)s:%(lineno)d]: %(message)s"
 )
