+++ src/red_log/loggers/__init__.py
@@ -1,1 +1,2 @@
 from ._loggers import LoggerConfig
+from ._factory import LoggerFactory
